Refactor MyPostsPage to use MyPostsParamsVerifier for handling post creation logic
