+++ src/app/my-posts/page.tsx
@@ -1,26 +1,14 @@
 'use client'
 
-import { useNewPost } from '@/features/posts/hooks/use-new-post'
-import { useSearchParams } from 'next/navigation'
-import { useEffect, useState } from 'react'
+import { MyPostsParamsVerifier } from '@/components/my-posts-params-verfier'
+import { Suspense } from 'react'
 
 export default function MyPostsPage() {
-  const [isLoaded, setIsLoaded] = useState<boolean>(false)
-  const { onOpen } = useNewPost()
-  const searchParams = useSearchParams()
-
-  useEffect(() => {
-    setIsLoaded(true)
-    if (isLoaded) {
-      const create = searchParams.get('create')
-      if (create === 'true') {
-        onOpen()
-      }
-    }
-  }, [searchParams, onOpen, isLoaded])
-
   return (
     <div className="container mx-auto px-4 py-8">
+      <Suspense>
+        <MyPostsParamsVerifier />
+      </Suspense>
       <div className="grid grid-cols-1 md:grid-cols-3 gap-6"></div>
     </div>
   )
